+++ Nodsoft.WowsReplaysUnpack.Core/Network/Packets/CellPlayerCreatePacket.cs
@@ -4,11 +4,11 @@ namespace Nodsoft.WowsReplaysUnpack.Core.Network.Packets;
 
 public class CellPlayerCreatePacket : ANetworkPacket
 {
-	public int EntityId { get; }
+	public uint EntityId { get; }
 	public byte[] Data { get; }
 	public CellPlayerCreatePacket(int packetIndex, BinaryReader binaryReader) : base(packetIndex)
 	{
-		EntityId = binaryReader.ReadInt32();
+		EntityId = (uint)binaryReader.ReadInt32();
 		_ = binaryReader.ReadInt32(); // spaceId;
 		_ = binaryReader.ReadInt32(); // vehicleId;
 		_ = binaryReader.ReadVector3(); // position (Vector3, 3xfloat)
