+++ Nodsoft.WowsReplaysUnpack.Core/Network/Packets/EntityControlPacket.cs
@@ -2,12 +2,12 @@
 
 public class EntityControlPacket : ANetworkPacket
 {
-	public int EntityId { get; }
+	public uint EntityId { get; }
 	public bool IsControlled { get; }
 
 	public EntityControlPacket(int packetIndex, BinaryReader binaryReader) : base(packetIndex)
 	{
-		EntityId = binaryReader.ReadInt32();
+		EntityId = (uint)binaryReader.ReadInt32();
 		IsControlled = binaryReader.ReadBoolean();
 	}
