+++ Nodsoft.WowsReplaysUnpack.Core/Network/Packets/EntityCreatePacket.cs
@@ -4,7 +4,7 @@ namespace Nodsoft.WowsReplaysUnpack.Core.Network.Packets;
 
 public class EntityCreatePacket : ANetworkPacket
 {
-	public int EntityId { get; }
+	public uint EntityId { get; }
 	public short Type { get; }
 	public int VehicleId { get; }
 	public int SpaceId { get; }
@@ -12,7 +12,7 @@ public class EntityCreatePacket : ANetworkPacket
 
 	public EntityCreatePacket(int packetIndex, BinaryReader binaryReader) : base(packetIndex)
 	{
-		EntityId = binaryReader.ReadInt32();
+		EntityId = (uint)binaryReader.ReadInt32();
 		Type = binaryReader.ReadInt16();
 		VehicleId = binaryReader.ReadInt32();
 		SpaceId = binaryReader.ReadInt32();
